Move DefaultFileMetadata to :files
