+++ subprojects/native/src/jmh/java/org/gradle/internal/nativeintegration/filesystem/FileMetadataAccessorBenchmark.java
@@ -17,6 +17,7 @@ package org.gradle.internal.nativeintegration.filesystem;
 
 import com.google.common.collect.ImmutableMap;
 import net.rubygrapefruit.platform.file.Files;
+import org.gradle.internal.file.DefaultFileMetadata;
 import org.gradle.internal.file.FileMetadataSnapshot;
 import org.gradle.internal.file.FileType;
 import org.gradle.internal.nativeintegration.filesystem.jdk7.Jdk7FileMetadataAccessor;
